Split out model from Entity Details template
