+++ src/components/ChipGroup/ChipGroup.tsx
@@ -9,21 +9,23 @@ type Props = {
   chips: Chip[];
 };
 
-const ChipGroup = ({ chips }: Props) => (
-  <div className="chip-group">
-    {chips &&
-      Object.values(chips).map(({ label, count }) => {
-        return (
-          count > 0 && (
-            <div className="p-chip" key={label}>
-              <span
-                className={`status-icon is-${label.toLowerCase()}`}
-              >{`${count} ${label}`}</span>
-            </div>
-          )
-        );
-      })}
-  </div>
-);
+const ChipGroup = ({ chips }: Props) => {
+  return (
+    <div className="chip-group">
+      {chips &&
+        Object.values(chips).map(({ label, count }) => {
+          return (
+            count > 0 && (
+              <div className="p-chip" key={label}>
+                <span
+                  className={`status-icon is-${label.toLowerCase()}`}
+                >{`${count} ${label}`}</span>
+              </div>
+            )
+          );
+        })}
+    </div>
+  );
+};
 
 export default ChipGroup;
