+++ src/declarations/global.d.ts
@@ -1,2 +1,4 @@
 declare module "enzyme";
 declare module "@canonical/react-components/dist/components/Modal";
+declare module "@canonical/react-components/dist/components/MainTable";
+
