+++ src/pages/EntityDetails/App/App.tsx
@@ -1,0 +1,7 @@
+type Props = {
+  name: string;
+};
+
+export default function App({ name }: Props) {
+  return <div>{name}</div>;
+}
