+++ src/pages/EntityDetails/counts.js
@@ -1,0 +1,35 @@
+export const incrementCounts = (status, counts) => {
+  if (counts[status]) {
+    counts[status] = counts[status] += 1;
+  } else {
+    counts[status] = 1;
+  }
+  return counts;
+};
+
+export const formatCounts = (counts) =>
+  Object.entries(counts).map((statusSet) => ({
+    count: statusSet[1],
+    label: statusSet[0],
+  }));
+
+export const generateSecondaryCounts = (modelStatusData, segment, selector) =>
+  Object.entries(modelStatusData[segment]).reduce((counts, section) => {
+    const status = section[1][selector].status;
+    return incrementCounts(status, counts);
+  }, {});
+
+export const generateUnitSecondaryCounts = (modelStatusData) => {
+  const counts = {};
+  let totalUnits = 0;
+  const applications = modelStatusData.applications;
+  Object.keys(applications).forEach((applicationName) => {
+    const units = applications[applicationName].units || [];
+    Object.keys(units).forEach((unitId) => {
+      const status = units[unitId]["agent-status"].status;
+      totalUnits += 1;
+      return incrementCounts(status, counts);
+    });
+  });
+  return [formatCounts(counts), totalUnits];
+};
