+++ src/pages/Models/_models.scss
@@ -13,7 +13,14 @@
     align-items: center;
     display: grid;
     grid-template-columns: 1fr;
+    opacity: 1;
     padding: 0.15rem;
+    transition: opacity 0.5s;
+
+    &[data-disabled="true"] {
+      opacity: 0.5;
+      pointer-events: none;
+    }
 
     .models__count {
       display: none;
